useMemo for defaultBarName in NewContainer
